requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.2.0
et-xmlfile==1.1.0
flake8==6.0.0
flake8-isort==6.0.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
numpy==1.25.1
openpyxl==3.1.2
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3
requests==2.31.0
six==1.16.0
tornado==6.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
